--- lamindb/models/_feature_constants.py
@@ -1,4 +1,0 @@
-"""Shared constants for feature metadata handling."""
-
-# Number of schema members included when constructing feature preview metadata.
-SCHEMA_MEMBER_PREVIEW_LIMIT = 20
